groups: never +go-restart-updates on group hosts
If the host finds itself in an inconsistent state where it can't apply
its own updates... In the abstract that shouldn't be possible. If we hit
that in practice, there's no way for us to recover, since we ourselves
are the source of truth. In light of that, probably better to hard-crash
than to try and do reconiliation that will never achieve anything.