groups: give single %create update in initial log
Instead of giving the full log and making the client replay it to arrive
at the correct latest state, we simply construct a faux %create update
containing the group as it exists at this point in time, and give that
as the whole log.

To make sure permission-related state gets updated appropriately, we
make channels-server recheck permissions on %create events for groups,
in addition to all the permission-related updates.