channels: re-start sequence number fetching flows
During the 8-to-9 migration, we would set timers for sequence number
fetching. However, due to poor implementation at the time, handling
those timers could crash. (Negotiate would detect pokes to mismatching
gills and rightfully abort the event.) That would result in the timer
for the affected channel being dropped on the floor, and sequence
numbers never getting fetched.

Now that the timer handling logic has been patched (70eae57) we should
make sure we pick up all the sequence numbers that we missed in this
way. Since we don't have easy/reliable access to information on which
channels were affected, we indiscriminately retry for _all_ channels.

If that results in duplicate timers, that's not the end of the world.
It's wasteful, but shouldn't give any real trouble.