Rename splash screen background color set to match Expo supported color
This flash was being caused by the iOS native root view always having a
white background color - in the frame where RN is not yet loaded, we see
that white view.

expo-updates recreates the root view after `[ExpoAppDelegate
customizeRootView:]` and assigns its own background color:
https://github.com/expo/expo/blame/63e051de05bc62bfad0c11dea0c64d2e750051e4/packages/expo-updates/ios/EXUpdates/ReactDelegateHandler/ExpoUpdatesReactDelegateHandler.swift#L48

This means we can't use customizeRootView: to reliably set the
background color to what we want.

That `ExpoUpdatesReactDelegateHandler` code attempts to set the
background color to the named color set `SplashScreenBackground` - this
isn't really documented (there's reference to this name in a
expo-splash-screen guide but seems more like a suggestion in a managed
app context?), but we can use this named color set to pass our own
colors to expo-updates, and make the native root view's background what
we want.