lure: keep lure link metadata in sync
Once a lure link has been generated, it is never updated.
This can causes various UX problems. Most recently group invites would
display with a generic "New Group" title.

This commit sets up a sync mechanism to update relevant lure links
both locally and at the bait provider on changes to user contact profile
or group metadata.

To minimize the bandwidth for groups with a lot of members, only
the group host triggers the bait provider to update lure invite metadata.
The facilitate this, the bait provider now keeps an index of group IDs to
the set associated lure invites.