channels/server: drop likely-bad (image) links
We may have raw image data links in our state. These bog down response
serialization and should never have existed.

Here we trawl through posts and logs in state, and drop any links in the
content that look out of the ordinary.

We opt not to issue any updates from the server, or even bump revision
numbers on affected posts. Clients will be dropping these links on their
side too, and formally syncing these changes complicates the logic and
increases risk on rollout. As written, this is a simple state change.

We take a somewhat broad definition of what a "good link" is. Anything
that starts with `http`, `#` or `/` is fair game. Note that this will
drop `mailto:` links and other such constructions. We may or may not
want to tighten this heuristic to only drop bad links beyond a certain
size.