groups: make %group-join poke more robust
Old %group-join poke was implemented in a way that required
an entry in foreigns to work. This was never a problem at the client,
because it would always preview groups before joining. However,
it caused crashes at hosting, who were issuing group join pokes to
public groups without previewing first.

We improve this by defaulting to a join without token when a foreign
group entry is not found.