logs: filter posthog logs
By default, only send logs with volume %info and higher into posthog,
dropping the remainder on the floor.

This is a partial revert of #5052's 89761bd. In that PR we introduced
additional %dbug logs but ended up not including log submission control.
The result is excessive log submission to posthog.

Here we reintroduce that log submission control.

Co-authored-by: Hunter Miller <hunter@hmiller.dev>