channels: include joins and leaves in /changes
Updates the agent-wide /changes endpoint to include channels when they
are created/joined or deleted/left. The former will include an empty
list of messages (unless there's messages since the change timestamp),
the latter will include a `null` entry for that channel.

We track a `last-updated` list in state so we can reflect this
accurately, taking care to update that whenever a channels gets created
or deleted, and whenever we receive any update for it.