chat: include joins and leaves in /changes
As the prior commit does for channels, so we do for chat here: change
the type of %chat-changed-writs and include entries if a chat was merely
created or deleted, with blank list of messages or `null` respectively.

Here, too, we assume the client wasn't using this endpoint yet and just
change the mark type.