channels/server: drop bad links from .meta too
In #5119 we had cleaned up "bad" links (usually: base64 blobs) from post
and log contents. However, we neglected to check the meta.essay, which
may have similar troubles in its image and cover fields.

We lightly restructure the existing utilities, to be a core with
multiple entry-points based on what type you want to drop bad links
from.

We update the callsites as appropriate, and _move_ the cleanup logic
into a new state transition to avoid duplicating effort for
not-yet-up-to-date ships.

Lastly we update the tests to match.