channels: ditch semi-redundant faces in $v-channel
$v-channel's $global type +rev'isions most of its child nouns. As
written, they get two faces: one on the outer [revision value] pair,
and one on the value itself.

The way that's written makes it easier to skip past the revision and get
just the value out, at the cost of making usage of the stuff _inside_
the value more verbose.

This commit makes the opposite tradeoff and removes the additional face.
This way, skipping past the revision to the raw value requires raw tree
addressing, but working with the stuff inside the value takes one less
face.

Presented as RFC. Made a note of this a while ago and wanted to try it
the other way, I'd be fine with either merging or closing. Argument in
favor of this is that usage happens spread out all over the place, but
stripping the revision off happens in more or less the same place for
all these types, limiting the cognitive drawback of looking at the lark.