metagrab: include accept-language header
If the request coming in to metagrab includes an accept-language header,
pass that on to whatever resource we're trying to grab metadata from.
This way, we are hopefully more likely to get results in the user's
language.

We restructure lightly to make request header tracking more generic and
require smaller diffs if we want to include another header in the
future. Instead of putting the user-agent header contents in the wire,
we put a %blob $coin containing the whole $header-list:http. We unpack
this carefully, making sure it fits the expected type, and falling back
to not retrieving headers instead of crashing the processing entirely.
(This is a stylistic/defensive choice. We will get out what we put in
100% of the time.)