scroller: switch to useLayoutEffect for sticky scrolling
Me and @davidisaaclee were looking at the scroller issue earlier today and while
we don't have a complete understanding of what exactly is going on here we do
know that this simple change fixes the "unstuck" scroller issue for all
scenarios we are able to reproduce.

This PR also reverts the 100px workaround that went live yesterday since it's
unnecessary now. This also means that the UX makes more sense again, and you're
not "stuck" when you've scrolled 99px up from the bottom.

Observations from @davidisaaclee:

```
1. scroll content size changes, which naturally triggers a useStickToStart scrollTo (because shouldStickToBottom is correctly true)
2. scrollTo triggers a scroll event, which is picked up by useScrollBoundary
3. concurrently*, the scroll's content height changes (likely due to "sending" arrows disappearing)
this triggers a ResizeObserver callback, which sets scrollerContentsKey state - but via a React state, which only enqueues an update to be run on next render
notably, scroller.scrollHeight will report the new height
4. useScrollBoundary looks at the new scroller.scrollHeight and sees that the scroll is no longer at the bottom, and sets insideBoundary to false
5. we're now unstuck
```