verb: intercept and print logs cards
We have hitherto printed logs of interest in each agent separately.
Recently, another approach had been tried of printing the received
logs in the logs agent. However, we realized that this approach
breaks down in some edge cases, such as the original event failing *before* the logging card reaches the logs agent.

Here we take another approach of printing the logging cards
in the verb library. Since lib-verb is an agent wrapper, the logging
message is sure to appear as soon as the card is emitted, even
if the event were to crash later.