notify: avoid "giving on system duct"
We dynamically find subscription paths to give facts on, but these lists
could be empty. If they're empty, we shouldn't give a fact: gall would
interpret it as giving a fact up the call stack, even though this isn't
an +on-watch scenario.

Additionally, make sure we don't crash when moving past empty paths.