groups: no-op on old-style channel sub sign
This code was already broken: in the "channel unknown" case it would
crash, and in the known channel case it would recur with a modified wire
that doesn't get caught by the `?+` and so would also crash.

Here, we turn the whole block into a no-op.

We were likely seeing this codepath get hit (during the current/upcoming
groups protocol update) because it's perfectly possible for preview
subscriptions to stay open indefinitely if the host isn't online to
handle our subscription request. The upgrade would kick those subs,
hitting this codepath.