api: when parsing `@ud`, also accept plain decimal
Presumably, the `parseUd()` of the old aura-js silently accepted decimal
number strings without dots, because the new `parse('ud', myString)`
calls were (in some cases) (we assume) crashing, failing to parse their
inputs.

Here, we add a helper that tries both and update all `parse('ud', ..)`
calls to use that instead. This may be more lenient than it strictly
needs to be, but we intentionally punt on unraveling that particularly
messy tapestry. (See also `time.md`...)